+++ src/components/common/index.ts
@@ -4,4 +4,3 @@ export { Select } from './Select';
 export { DatePicker } from './DatePicker';
 export { TextArea } from './TextArea';
 export { ProtectedRoute } from './ProtectedRoute';
-export { Menubar } from './Menubar';
